+++ test/utils.system.js
@@ -22,4 +22,16 @@ describe("Utils System Tests", function() {
             done("Could not get CPU core count from server info result.")
         }
     })
+
+    it("Get server info without labels", function(done) {
+        var serverInfo = utils.system.getInfo({
+            labels: false
+        })
+
+        if (serverInfo.memoryUsage.toString().indexOf("%") > 0 || serverInfo.memoryTotal.toString().indexOf("MB") > 0) {
+            done("Output should not include labels % MB etc.")
+        } else {
+            done()
+        }
+    })
 })